@@ -92,12 +92,12 @@
     "babel-loader": "9.2.1",
     "base64-loader": "1.0.0",
     "browserslist": "4.23.2",
-    "chromatic": "13.1.2",
+    "chromatic": "13.3.1",
     "concurrently": "9.2.0",
     "copy-webpack-plugin": "13.0.0",
     "cross-env": "10.1.0",
     "css-loader": "7.1.2",
-    "electron": "36.9.3",
+    "electron": "37.7.0",
     "electron-builder": "26.0.12",
     "electron-log": "5.4.0",
     "electron-reload": "2.0.0-alpha.1",
@@ -160,8 +160,8 @@
     "@angular/platform-browser": "19.2.14",
     "@angular/platform-browser-dynamic": "19.2.14",
     "@angular/router": "19.2.14",
-    "@bitwarden/sdk-internal": "0.2.0-main.357",
-    "@bitwarden/commercial-sdk-internal": "0.2.0-main.357",
+    "@bitwarden/sdk-internal": "0.2.0-main.365",
+    "@bitwarden/commercial-sdk-internal": "0.2.0-main.365",
     "@electron/fuses": "1.8.0",
     "@emotion/css": "11.13.5",
     "@koa/multer": "4.0.0",
@@ -193,7 +193,7 @@
     "lowdb": "1.0.0",
     "lunr": "2.3.9",
     "multer": "2.0.2",
-    "ngx-toastr": "19.0.0",
+    "ngx-toastr": "19.1.0",
     "node-fetch": "2.6.12",
     "node-forge": "1.3.1",
     "oidc-client-ts": "2.4.1",
