@@ -96,7 +96,7 @@ impl MacOSProviderClient {
             response_callbacks_queue: Arc::new(Mutex::new(HashMap::new())),
         };
 
-        let path = desktop_core::ipc::path("autofill");
+        let path = desktop_core::ipc::path("af");
 
         let queue = client.response_callbacks_queue.clone();
 
