@@ -5,9 +5,9 @@ import { ErrorResponse } from "@bitwarden/common/models/response/error.response"
 import { OrganizationId, OrganizationReportId } from "@bitwarden/common/types/guid";
 
 import {
-  EncryptedDataWithKey,
   UpdateRiskInsightsApplicationDataRequest,
   UpdateRiskInsightsApplicationDataResponse,
+  UpdateRiskInsightsSummaryDataRequest,
 } from "../../models";
 import {
   GetRiskInsightsApplicationDataResponse,
@@ -29,7 +29,9 @@ export class RiskInsightsApiService {
       true,
     );
     return from(dbResponse).pipe(
-      map((response) => new GetRiskInsightsReportResponse(response)),
+      // As of this change, the server doesn't return a 404 if a report is not found
+      // Handle null response if server returns nothing
+      map((response) => (response ? new GetRiskInsightsReportResponse(response) : null)),
       catchError((error: unknown) => {
         if (error instanceof ErrorResponse && error.statusCode === 404) {
           return of(null); // Handle 404 by returning null or an appropriate default value
@@ -73,14 +75,14 @@ export class RiskInsightsApiService {
   }
 
   updateRiskInsightsSummary$(
-    summaryData: EncryptedDataWithKey,
-    organizationId: OrganizationId,
     reportId: OrganizationReportId,
+    organizationId: OrganizationId,
+    request: UpdateRiskInsightsSummaryDataRequest,
   ): Observable<void> {
     const dbResponse = this.apiService.send(
       "PATCH",
       `/reports/organizations/${organizationId.toString()}/data/summary/${reportId.toString()}`,
-      summaryData,
+      { ...request.data, reportId: reportId, organizationId },
       true,
       true,
     );
