@@ -1,3 +1,4 @@
+import { takeUntilDestroyed } from "@angular/core/rxjs-interop";
 import { BehaviorSubject } from "rxjs";
 
 import { ApplicationHealthReportDetailEnriched } from "../../models";
@@ -20,7 +21,6 @@ export class AllActivitiesService {
     totalCriticalApplicationCount: 0,
     totalAtRiskApplicationCount: 0,
     totalCriticalAtRiskApplicationCount: 0,
-    newApplications: [],
   });
   reportSummary$ = this.reportSummarySubject$.asObservable();
 
@@ -40,14 +40,15 @@ export class AllActivitiesService {
 
   constructor(private dataService: RiskInsightsDataService) {
     // All application summary changes
-    this.dataService.enrichedReportData$.subscribe((report) => {
+    this.dataService.enrichedReportData$.pipe(takeUntilDestroyed()).subscribe((report) => {
       if (report) {
         this.setAllAppsReportSummary(report.summaryData);
         this.setAllAppsReportDetails(report.reportData);
       }
     });
+
     // Critical application summary changes
-    this.dataService.criticalReportResults$.subscribe((report) => {
+    this.dataService.criticalReportResults$.pipe(takeUntilDestroyed()).subscribe((report) => {
       if (report) {
         this.setCriticalAppsReportSummary(report.summaryData);
       }
@@ -55,6 +56,9 @@ export class AllActivitiesService {
   }
 
   setCriticalAppsReportSummary(summary: OrganizationReportSummary) {
+    if (!summary) {
+      return;
+    }
     this.reportSummarySubject$.next({
       ...this.reportSummarySubject$.getValue(),
       totalCriticalApplicationCount: summary.totalApplicationCount,
@@ -65,13 +69,16 @@ export class AllActivitiesService {
   }
 
   setAllAppsReportSummary(summary: OrganizationReportSummary) {
+    if (!summary) {
+      return;
+    }
+
     this.reportSummarySubject$.next({
       ...this.reportSummarySubject$.getValue(),
       totalMemberCount: summary.totalMemberCount,
       totalAtRiskMemberCount: summary.totalAtRiskMemberCount,
       totalApplicationCount: summary.totalApplicationCount,
       totalAtRiskApplicationCount: summary.totalAtRiskApplicationCount,
-      newApplications: summary.newApplications,
     });
   }
 
