@@ -4,7 +4,13 @@ import { distinctUntilChanged, map } from "rxjs/operators";
 import { OrganizationId } from "@bitwarden/common/types/guid";
 
 import { getAtRiskApplicationList, getAtRiskMemberList } from "../../helpers";
-import { ReportState, DrawerDetails, DrawerType, RiskInsightsEnrichedData } from "../../models";
+import {
+  ReportState,
+  DrawerDetails,
+  DrawerType,
+  RiskInsightsEnrichedData,
+  ReportStatus,
+} from "../../models";
 import { RiskInsightsOrchestratorService } from "../domain/risk-insights-orchestrator.service";
 
 export class RiskInsightsDataService {
@@ -24,10 +30,15 @@ export class RiskInsightsDataService {
   // -------------------------- Orchestrator-driven state  -------------
   // The full report state (for internal facade use or complex components)
   private readonly reportState$: Observable<ReportState>;
-  readonly isLoading$: Observable<boolean> = of(false);
+  readonly reportStatus$: Observable<ReportStatus> = of(ReportStatus.Initializing);
+  readonly hasReportData$: Observable<boolean> = of(false);
   readonly enrichedReportData$: Observable<RiskInsightsEnrichedData | null> = of(null);
   readonly isGeneratingReport$: Observable<boolean> = of(false);
   readonly criticalReportResults$: Observable<RiskInsightsEnrichedData | null> = of(null);
+  readonly hasCiphers$: Observable<boolean | null> = of(null);
+
+  // New applications that need review (reviewedDate === null)
+  readonly newApplications$: Observable<string[]> = of([]);
 
   // ------------------------- Drawer Variables ---------------------
   // Drawer variables unified into a single BehaviorSubject
@@ -48,11 +59,18 @@ export class RiskInsightsDataService {
     this.organizationDetails$ = this.orchestrator.organizationDetails$;
     this.enrichedReportData$ = this.orchestrator.enrichedReportData$;
     this.criticalReportResults$ = this.orchestrator.criticalReportResults$;
+    this.newApplications$ = this.orchestrator.newApplications$;
+
+    this.hasCiphers$ = this.orchestrator.hasCiphers$.pipe(distinctUntilChanged());
 
     // Expose the loading state
-    this.isLoading$ = this.reportState$.pipe(
-      map((state) => state.loading),
+    this.reportStatus$ = this.reportState$.pipe(
+      map((state) => state.status),
       distinctUntilChanged(), // Prevent unnecessary component re-renders
+    );
+    this.hasReportData$ = this.reportState$.pipe(
+      map((state) => state.data != null),
+      distinctUntilChanged(),
     );
   }
 
@@ -68,10 +86,6 @@ export class RiskInsightsDataService {
 
   triggerReport(): void {
     this.orchestrator.generateReport();
-  }
-
-  fetchReport(): void {
-    this.orchestrator.fetchReport();
   }
 
   // ------------------------- Drawer functions -----------------------------
@@ -241,5 +255,9 @@ export class RiskInsightsDataService {
 
   removeCriticalApplication(hostname: string) {
     return this.orchestrator.removeCriticalApplication$(hostname);
+  }
+
+  saveApplicationReviewStatus(selectedCriticalApps: string[]) {
+    return this.orchestrator.saveApplicationReviewStatus$(selectedCriticalApps);
   }
 }
