@@ -1,4 +1,4 @@
-import { firstValueFrom } from "rxjs";
+import { firstValueFrom, Observable, of } from "rxjs";
 
 import { AccountService } from "@bitwarden/common/auth/abstractions/account.service";
 import { getUserId } from "@bitwarden/common/auth/services/account.service";
@@ -9,7 +9,12 @@ import { ExportedVault } from "../types";
 
 import { IndividualVaultExportServiceAbstraction } from "./individual-vault-export.service.abstraction";
 import { OrganizationVaultExportServiceAbstraction } from "./org-vault-export.service.abstraction";
-import { ExportFormat, VaultExportServiceAbstraction } from "./vault-export.service.abstraction";
+import {
+  ExportFormat,
+  ExportFormatMetadata,
+  FormatOptions,
+  VaultExportServiceAbstraction,
+} from "./vault-export.service.abstraction";
 
 export class VaultExportService implements VaultExportServiceAbstraction {
   constructor(
@@ -83,6 +88,26 @@ export class VaultExportService implements VaultExportServiceAbstraction {
       format,
       onlyManagedCollections,
     );
+  }
+
+  /**
+   * Get available export formats based on vault context
+   * @param options Options determining which formats are available
+   * @returns Observable stream of available export formats
+   */
+  formats$(options: FormatOptions): Observable<ExportFormatMetadata[]> {
+    const baseFormats: ExportFormatMetadata[] = [
+      { name: ".json", format: "json" },
+      { name: ".csv", format: "csv" },
+      { name: ".json (Encrypted)", format: "encrypted_json" },
+    ];
+
+    // ZIP format with attachments is only available for individual vault exports
+    if (options.isMyVault) {
+      return of([...baseFormats, { name: ".zip (with attachments)", format: "zip" }]);
+    }
+
+    return of(baseFormats);
   }
 
   /** Checks if the provided userId matches the currently authenticated user
